Access semantic CSS from fomantic package via CDN
The fomantic-ui-css NPM package is a community maintained fork of
semantic-ui-css, which itself is no longer maintained. There is
currently a bug with webpack and the semantic CSS file that is installed
locally by NPM (i.e. into node_modules/) that prevents compiling.
Accessing the semantic CSS file from the fomantic NPM package via the
CDN seems to work.